requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
flask-swagger-ui==4.11.1
h11==0.14.0
idna==3.7
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
outcome==1.3.0.post0
packaging==24.0
PyJWT==2.8.0
PySocks==1.7.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
undetected-chromedriver==3.5.5
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
